extract OnboardingGradient
